Group more datastore-related services in namespace
Grouping all service objects related to the datastore under the same namespace.
