+++ app/services/decisions/submission_decision_tree.rb
@@ -17,7 +17,7 @@ module Decisions
       # Get it before we purge the local DB record
       reference = current_crime_application.usn
 
-      if ApplicationSubmission.new(current_crime_application).call
+      if Datastore::ApplicationSubmission.new(current_crime_application).call
         show(:confirmation, reference:)
       else
         edit(:failure)
